More sign fixes

+++ modules/pico_ipv6.c
@@ -171,7 +171,7 @@ int pico_ipv6_to_string(char *ipbuf, const uint8_t ip[PICO_SIZE_IP6])
     }
 
     /* every nibble is one char */
-    for (i = 0; i < PICO_SIZE_IP6 * 2; ++i) {
+    for (i = 0; i < PICO_SIZE_IP6 * 2u; ++i) {
         if (i % 4 == 0 && i != 0)
             *ipbuf++ = ':';
 
@@ -906,7 +906,7 @@ void pico_ipv6_nd_dad(unsigned long now, void *arg)
     if (l->isduplicate) {
         dbg("IPv6: duplicate address.\n");
         if (pico_ipv6_is_linklocal(address.addr)) {
-            address.addr[8] = ((pico_rand() & 0xff) & (uint8_t)(~0x03));
+            address.addr[8] = ((uint8_t)(pico_rand() & 0xff) & (uint8_t)(~0x03));
             address.addr[9] = pico_rand() & 0xff;
             address.addr[10] = pico_rand() & 0xff;
             address.addr[11] = pico_rand() & 0xff;
